@@ -27,26 +27,26 @@ dependencies:
     sdk: flutter
   auto_size_text: 3.0.0
   cached_network_image: 3.2.1
-  dropdown_button2: 2.0.0
+  dropdown_button2: 2.3.9
   flutter_animate: 4.1.1+1
-  flutter_cache_manager: 3.3.0
-  font_awesome_flutter: 10.1.0
+  flutter_cache_manager: 3.3.1
+  font_awesome_flutter: 10.6.0
   from_css_color: 2.0.0
   go_router: 7.1.1
   google_fonts: 4.0.3
   intl: 0.18.1
   json_path: 0.6.2
-  page_transition: 2.0.4
+  page_transition: 2.1.0
   path_provider: 2.0.14
   path_provider_android: 2.0.25
   path_provider_foundation: 2.2.2
   path_provider_platform_interface: 2.0.6
-  provider: 6.0.4
-  shared_preferences: 2.0.15
-  shared_preferences_android: 2.1.0
-  shared_preferences_ios: 2.1.1
-  shared_preferences_platform_interface: 2.2.0
-  shared_preferences_web: 2.1.0
+  provider: 6.0.5
+  shared_preferences: 2.2.2
+  shared_preferences_android: 2.2.1
+  shared_preferences_foundation: 2.3.4
+  shared_preferences_platform_interface: 2.3.1
+  shared_preferences_web: 2.2.1
   sqflite: 2.2.6
   timeago: 3.2.2
   url_launcher: 6.1.10
