@@ -2,7 +2,7 @@
 export '/pages/home_page/home_page_widget.dart' show HomePageWidget;
 export '/pages/choose_eating_drinking/choose_eating_drinking_widget.dart'
     show ChooseEatingDrinkingWidget;
-export '/pages/eating_page/eating_page_widget.dart' show EatingPageWidget;
-export '/pages/drinking_page/drinking_page_widget.dart' show DrinkingPageWidget;
+export '/pages/orders_page/orders_page_widget.dart' show OrdersPageWidget;
+export '/pages/bill_page/bill_page_widget.dart' show BillPageWidget;
 export '/pages/choose_eating_order_page/choose_eating_order_page_widget.dart'
     show ChooseEatingOrderPageWidget;
